Enable CLI.
